pyproject.toml
[project]
name = "prompt-optimizer-proxy"
version = "0.1.0"
description = "A proxy that rewrites prompts before calling OpenAI (Responses API)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "openai>=1.45.0",
  "httpx>=0.27",
  "tenacity>=9.0",
  "pytest>=8.3",
]

[tool.uvicorn]
reload = true
host = "127.0.0.1"
port = 8000
